refactor: edit constants
